feat(scene.py): add `PkScene` class
A scene is a state of the game, e.g. the main menu, the game itself, the settings menu, etc. Each scene has its own class, which inherits from this class. The scenes are stored in the `scenes` dict in the `PkApp` class. A scene takes up the whole screen (minus bars).